requirements.txt
langchain
langchain-groq
streamlit
DateTime
python-docx
orjson